warp-terminal: add wayland to runtimeDeps by default
There isn't really a point to make "waylandSupport optional".
In 2025, you'd have to try really hard to have setup without Wayland.
You probably have it in your closure already.
Wayland is the present, and is here to stay for the forseeable future.
We already include a lot of Xorg stuff on Wayland setups anyway, that probably shouldn't be included.

For end users, what it changes is: it removes the waylandSupport override flag
You no longer need to set it, just set `WARP_ENABLE_WAYLAND=1` in your environment
via sessionVariables, or via home-manager or manually if that's what you wish.

Closes https://github.com/NixOS/nixpkgs/issues/393675

Signed-off-by: John Titor <50095635+JohnRTitor@users.noreply.github.com>